Retarget SDK to upstream Go toolchain 1.24 (#1)
* bugfix: fix Errorf without a format string


* feat: rework API to support full upstream Go compiler v1.24


* chore: byte slice conversion functions to stdlib implementations


* Revert "more updates"

This reverts commit 33c01aaf132cc5f82215bf5bba24fa0c2ec3743a.


* Revert "Deletes unnecessary components: part 2  (#453)"

This reverts commit f4829de2e48984b93c70b7d12ce9ebbbf955cc9d.


* Revert "Deletes unnecessary components (#452)"

This reverts commit cbcb26093a16604f5f62c587ed24ffa512c68ae0.


* Revert "Updates README to reflect the project direction (#451)"

This reverts commit 1d2249f168934bb9fa91a0fb4f6991f7d8ae5e40.


* owners: update for full Go version maintainer


* doc: update OVERVIEW for full Go compiler features and limitations


* doc: modernize README with full Go instructions


* chore: update issue template to ask which compiler is in use


* chore: update Makefile with full Go instructions


* chore: update workflows with full go sdk version 1.24


* chore: update required go.mod versions


* chore: fix lint errors


* doc: document alpha status in README


* chore: update references from tetratelabs repo to proxy-wasm repo


* doc: update docs with feedback on PR#1

- Add CLA and community guidelines to CONTRIBUTING.md
- Rename workflows
- Remove e2e tests from workflows
- Clarify difference from tetratelabs/proxy-wasm-go-sdk


* ci: rename workflow to "Presubmit checks"


* doc: update overview with review comments


* doc: replace tabs with spaces


* doc: replace tabs with spaces in README.md


* ci: fix workflow file syntax


* chore: fix import ordering post rename


* chore: add martijneken to CODEOWNERS


---------

+++ proxywasm/internal/timing_on.go
@@ -29,5 +29,5 @@ func logTiming(msg string, start time.Time) {
 		panic("BUG: logTiming should not be called when timing is disabled")
 	}
 	f := fmt.Sprintf("%s took %s", msg, time.Since(start))
-	ProxyLog(LogLevelDebug, StringBytePtr(f), len(f))
+	ProxyLog(LogLevelDebug, StringBytePtr(f), int32(len(f)))
 }
